优化接口

+++ src/larke/admin/Controller/AuthGroup.php
@@ -108,9 +108,8 @@ class AuthGroup extends Base
      * @param  Request  $request
      * @return Response
      */
-    public function detail(Request $request)
+    public function detail(string $id, Request $request)
     {
-        $id = $request->get('id');
         if (empty($id)) {
             return $this->errorJson(__('ID不能为空'));
         }
@@ -137,9 +136,8 @@ class AuthGroup extends Base
      * @param  Request  $request
      * @return Response
      */
-    public function delete(Request $request)
+    public function delete(string $id, Request $request)
     {
-        $id = $request->get('id');
         if (empty($id)) {
             return $this->errorJson(__('ID不能为空'));
         }
@@ -148,6 +146,10 @@ class AuthGroup extends Base
             ->first();
         if (empty($info)) {
             return $this->errorJson(__('信息不存在'));
+        }
+        
+        if ($info->is_system == 1) {
+            return $this->errorJson(__('系统信息不能删除'));
         }
         
         $deleteStatus = $info->delete();
@@ -210,9 +212,8 @@ class AuthGroup extends Base
      * @param  Request  $request
      * @return Response
      */
-    public function update(Request $request)
+    public function update(string $id, Request $request)
     {
-        $id = $request->get('id');
         if (empty($id)) {
             return $this->errorJson(__('ID不能为空'));
         }
@@ -264,9 +265,8 @@ class AuthGroup extends Base
      * @param  Request  $request
      * @return Response
      */
-    public function listorder(Request $request)
+    public function listorder(string $id, Request $request)
     {
-        $id = $request->get('id');
         if (empty($id)) {
             return $this->errorJson(__('ID不能为空'));
         }
@@ -295,9 +295,8 @@ class AuthGroup extends Base
      * @param  Request  $request
      * @return Response
      */
-    public function enable(Request $request)
+    public function enable(string $id, Request $request)
     {
-        $id = $request->get('id');
         if (empty($id)) {
             return $this->errorJson(__('ID不能为空'));
         }
@@ -326,9 +325,8 @@ class AuthGroup extends Base
      * @param  Request  $request
      * @return Response
      */
-    public function disable(Request $request)
+    public function disable(string $id, Request $request)
     {
-        $id = $request->get('id');
         if (empty($id)) {
             return $this->errorJson(__('ID不能为空'));
         }
@@ -357,9 +355,8 @@ class AuthGroup extends Base
      * @param  Request  $request
      * @return Response
      */
-    public function access(Request $request)
+    public function access(string $id, Request $request)
     {
-        $id = $request->get('id');
         if (empty($id)) {
             return $this->errorJson(__('ID不能为空'));
         }
