+++ src/larke/admin/Controller/AuthRule.php
@@ -90,7 +90,7 @@ class AuthRule extends Base
      */
     public function indexChildren(Request $request)
     {
-        $id = $request->get('id', 0);
+        $id = $request->get('id');
         if (empty($id) || is_array($id)) {
             return $this->errorJson(__('ID不能为空'));
         }
@@ -113,9 +113,8 @@ class AuthRule extends Base
      * @param  Request  $request
      * @return Response
      */
-    public function detail(Request $request)
+    public function detail(string $id, Request $request)
     {
-        $id = $request->get('id');
         if (empty($id)) {
             return $this->errorJson(__('ID不能为空'));
         }
@@ -135,9 +134,8 @@ class AuthRule extends Base
      * @param  Request  $request
      * @return Response
      */
-    public function delete(Request $request)
+    public function delete(string $id, Request $request)
     {
-        $id = $request->get('id');
         if (empty($id)) {
             return $this->errorJson(__('ID不能为空'));
         }
@@ -152,6 +150,10 @@ class AuthRule extends Base
             ->first();
         if (!empty($childInfo)) {
             return $this->errorJson(__('还有子权限链接存在，请删除子权限链接后再操作！'));
+        }
+        
+        if ($info->is_system == 1) {
+            return $this->errorJson(__('系统权限不能删除'));
         }
         
         $deleteStatus = $info->delete();
@@ -229,11 +231,8 @@ class AuthRule extends Base
      * @param  Request  $request
      * @return Response
      */
-    public function update(Request $request)
+    public function update(string $id, Request $request)
     {
-        $requestData = $request->all();
-        
-        $id = $request->get('id');
         if (empty($id)) {
             return $this->errorJson(__('账号ID不能为空'));
         }
@@ -244,6 +243,7 @@ class AuthRule extends Base
             return $this->errorJson(__('信息不存在'));
         }
         
+        $requestData = $request->all();
         $validator = Validator::make($requestData, [
             'parentid' => 'required',
             'title' => 'required|max:50',
@@ -306,9 +306,8 @@ class AuthRule extends Base
      * @param  Request  $request
      * @return Response
      */
-    public function listorder(Request $request)
+    public function listorder(string $id, Request $request)
     {
-        $id = $request->get('id');
         if (empty($id)) {
             return $this->errorJson(__('ID不能为空'));
         }
@@ -337,9 +336,8 @@ class AuthRule extends Base
      * @param  Request  $request
      * @return Response
      */
-    public function enable(Request $request)
+    public function enable(string $id, Request $request)
     {
-        $id = $request->get('id');
         if (empty($id)) {
             return $this->errorJson(__('ID不能为空'));
         }
@@ -368,9 +366,8 @@ class AuthRule extends Base
      * @param  Request  $request
      * @return Response
      */
-    public function disable(Request $request)
+    public function disable(string $id, Request $request)
     {
-        $id = $request->get('id');
         if (empty($id)) {
             return $this->errorJson(__('ID不能为空'));
         }
